XXXX Automated Testing using Jenkins and EC2
This adds the necessary files such that we can leverage Jenkins and
Amazon EC2 to perfrom basic sanity testing of OpenZFS pull requests.
With the included "Jenkinsfile", the following will be performed for
each pull request submitted:

  1. Create a new VM in Amazon EC2 based on OpenIndiana "Hipster"
  2. Configure the VM from (1) to support building OpenZFS
  3. Perform a full "nightly" build
  4. Run "git nits"
  5. Upgrade the VM from (1) using the build products from (3)
  6. Clone the upgraded VM from (5), and run the following tests (each
     test being run in parallel, and on an independent Amazon EC2 VM):
     "libctest", "ostest", "utiltest", "zfstest", and "zloop".

Additionally, the contents of "/var/tmp/test_result" and the log file
associated with each individual test will be archived and available for
download via the Jenkins UI; which is useful for debugging failed tests.