8125 kmem_move tunables must not be declared static
Reviewed by: Serapheim Dimitropoulos <serapheim@delphix.com>
Reviewed by: Dan Kimmel <dan.kimmel@delphix.com>
Reviewed by: Prashanth Sreenivasa <pks@delphix.com>
Reviewed by: Toomas Soome <tsoome@me.com>
Reviewed by: Igor Kozhukhov <igor@dilos.org>
Reviewed by: Robert Mustacchi <rm@joyent.com>

The kmem_move_* tunables (e.g. kmem_move_noreap) are declared "static".
Because they are never set in the code, the compiler can consider them
constants and optimize them away - removing the code that checks them.

They must not be declared "static" to be used as tunables.

Upstream bugs: DLPX-50335