WIP: ZIL Performance Improvements
This is still a work in progress, and isn't meant to be landed in its
current state. I'm only pushing this code publicly now, rather than when
it's "done", so the design of the changes implemented here can be
compared with the code in the ZFS on Linux PR here:
https://github.com/zfsonlinux/zfs/pull/6133

There's a "high level" comment about the new design of `zil_commit()` in
a code comment above that function; please see that for more details.