8484 Implement aggregate sum and use for arc counters
Reviewed by: Pavel Zakharov <pavel.zakharov@delphix.com>
Reviewed by: Matthew Ahrens <mahrens@delphix.com>

In pursuit of improving performance on multi-core systems, this patch
implements fanned out counters and uses them to improve the performance of
some of the arc statistics. These stats are updated extremely frequently,
and can consume a significant amount of CPU time.

The cityhash code is copied from google's github, where it is MIT licensed.

Upstream bugs: DLPX-51685