Archive the "mail_msg" file + "package" directory
In addition to archiving the build's "nightly.log" file, this change
adds support for archiving the build's "mail_msg" file and the
"packages" directory.

The contents of "mail_msg" are already visible in the pipeline view of
the build, but having it in the build's artifactsn makes it easier to
share via a direct link to the file via the build artifacts (e.g. for
illumos RTI purposes).

Also, by including the "packages" directory in the build's artifacts, it
makes it easier to stand up an OpenIndiana system running the packages
produced via the Jenkins build. For example, one can spin up a new
system (e.g. using the OpenZFS AMI), download the "nightly-packages"
artifact, and then use those packages to pass into "onu" to upgrade the
system. This can make it easier to do manual testing and verfication,
since one can rely on the Jenkins automation to perform the build.