Limit the uptime of Jenkins slaves to 12 hours
To try and prevent EC2 instance "leaks", this patch adds an additional
daemon that is managed by supervisord. The purpose of this new daemon
is to shutdown the system after it has been running for a specified
amount of time. As a result of the shutdown, the instance should be
automatically terminated by AWS since we're exclusively using EC2 Spot
Instances for all Jenkins slaves.