0000 ztest assertion failure in zil_lwb_write_issue
Reviewed by: Matt Ahrens <mahrens@delphix.com>
Reviewed by: Don Brady <don.brady@delphix.com>

When TXG_WAITED is passed in to dmu_tx_assign(), it's possible that
ERESTART will be returned, which can cause an assertion to fail in
zil_lwb_write_issue(). To fix this, dmu_tx_assign() was modifed to only
return ERESTART if TXG_NOWAIT is passed in.

Upstream bugs: DLPX-54873